requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
streamlit>=1.30.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
